+++ courier-compiler/src/main/java/me/denley/courier/compiler/DataMapProcessor.java
@@ -23,9 +23,11 @@ public class DataMapProcessor extends AbstractProcessor {
 
     public static final String CLASS_SUFFIX = "$$DataMapPackager";
 
+    private static final String BITMAP = "android.graphics.Bitmap";
+
     private static final String INDENT = "    ";
     private static final String INDENT_2 = "        ";
-
+    private static final String INDENT_3 = "            ";
 
     Set<String> targetClassNames = new LinkedHashSet<String>();
 
@@ -125,9 +127,17 @@ public class DataMapProcessor extends AbstractProcessor {
         final String parentClass = getParentClassName(element);
 
         builder.append("package ").append(getPackageName(element)).append(";\n\n");
+        builder.append("import me.denley.courier.Courier;\n");
         builder.append("import me.denley.courier.Packager;\n");
         builder.append("import me.denley.courier.Packager.DataPackager;\n");
-        builder.append("import com.google.android.gms.wearable.DataMap;\n\n");
+        builder.append("import com.google.android.gms.wearable.DataMap;\n");
+        builder.append("import com.google.android.gms.wearable.Asset;\n");
+        builder.append("import android.content.Context;\n");
+        builder.append("import android.graphics.Bitmap;\n");
+        builder.append("import android.graphics.BitmapFactory;\n");
+        builder.append("import java.io.InputStream;\n");
+        builder.append("import java.io.ByteArrayOutputStream;\n\n");
+
         writeClassDef(builder, element, parentClass);
     }
 
@@ -164,9 +174,9 @@ public class DataMapProcessor extends AbstractProcessor {
     }
 
     private void writeUnpackInterfaceMethod(StringBuilder builder, String targetClassName) {
-        builder.append(INDENT).append("public T unpack(DataMap map) {\n");
+        builder.append(INDENT).append("public T unpack(Context context, DataMap map) {\n");
         builder.append(INDENT_2).append("final T target = (T) new ").append(targetClassName).append("();\n");
-        builder.append(INDENT_2).append("unpack(map, target);\n");
+        builder.append(INDENT_2).append("unpack(context, map, target);\n");
         builder.append(INDENT_2).append("return target;\n");
         builder.append(INDENT).append("}\n\n");
     }
@@ -192,9 +202,17 @@ public class DataMapProcessor extends AbstractProcessor {
 
                 final DataMapElementType elementType = DataMapElementType.getElementType(subElement);
                 if(elementType==null) {
-                    if(targetClassNames.contains(subElement.asType().toString())) {
+                    final String fieldType = subElement.asType().toString();
+                    if(targetClassNames.contains(fieldType)) {
                         builder.append(INDENT_2).append("map.putDataMap(\"").append(name).append("\", ")
                                 .append("Packager.pack(target.").append(name).append("));\n");
+                    } else if (fieldType.equals(BITMAP)) {
+                        builder.append(INDENT_2).append("if(target.").append(name).append("!=null) {\n");
+                        builder.append(INDENT_3).append("final ByteArrayOutputStream ").append(name).append("ByteArrayOutputStream = new ByteArrayOutputStream();\n");
+                        builder.append(INDENT_3).append("target.").append(name).append(".compress(Bitmap.CompressFormat.PNG, 100, ").append(name).append("ByteArrayOutputStream);\n");
+                        builder.append(INDENT_3).append("final Asset ").append(name).append("Asset = Asset.createFromBytes(").append(name).append("ByteArrayOutputStream.toByteArray());\n");
+                        builder.append(INDENT_3).append("map.putAsset(\"").append(name).append("\", ").append(name).append("Asset);\n");
+                        builder.append(INDENT_2).append("}\n");
                     } else {
                         throw new IllegalArgumentException("Field type ("+subElement.asType().toString()+") is not mappable to a DataMap.");
                     }
@@ -212,9 +230,9 @@ public class DataMapProcessor extends AbstractProcessor {
         if(parentClass!=null) {
             builder.append(INDENT).append("@Override\n");
         }
-        builder.append(INDENT).append("protected void unpack(DataMap map, T target) {\n");
+        builder.append(INDENT).append("protected void unpack(Context context, DataMap map, T target) {\n");
         if(parentClass!=null) {
-            builder.append(INDENT_2).append("super.unpack(map, target);\n");
+            builder.append(INDENT_2).append("super.unpack(context, map, target);\n");
         }
 
         for(Element subElement:element.getEnclosedElements()) {
@@ -229,12 +247,19 @@ public class DataMapProcessor extends AbstractProcessor {
 
                 final DataMapElementType elementType = DataMapElementType.getElementType(subElement);
                 if(elementType==null) {
-                    if(targetClassNames.contains(subElement.asType().toString())) {
+                    final String fieldType = subElement.asType().toString();
+                    if(targetClassNames.contains(fieldType)) {
                         builder.append(INDENT_2).append("target.").append(name).append(" = ");
                         builder.append("Packager.unpack(")
                                 .append("map.getDataMap(\"").append(name)
                                 .append("\"), ")
                                 .append(subElement.asType().toString()).append(".class);\n");
+                    } else if (fieldType.equals(BITMAP)) {
+                        builder.append(INDENT_2).append("final Asset ").append(name).append("Asset = map.getAsset(\"").append(name).append("\");\n");
+                        builder.append(INDENT_2).append("if(").append(name).append("Asset!=null && context!=null) {\n");
+                        builder.append(INDENT_3).append("final InputStream in = Courier.getAssetInputStream(context, ").append(name).append("Asset);\n");
+                        builder.append(INDENT_3).append("target.").append(name).append(" = BitmapFactory.decodeStream(in);\n");
+                        builder.append(INDENT_2).append("}\n");
                     } else {
                         throw new IllegalArgumentException("Field type ("+subElement.asType().toString()+") is not mappable to a DataMap.");
                     }